Organize config into topical files
