+++ packages/web/config/ibc-assets.ts
@@ -1,0 +1,194 @@
+import { IBCAsset } from "../stores/assets";
+
+/**
+ * Determine the channel info per the chain.
+ * Guide users to use the same channel for convenience.
+ */
+export const IBCAssetInfos: IBCAsset[] = [
+  {
+    counterpartyChainId: "cosmoshub-4",
+    sourceChannelId: "channel-0",
+    destChannelId: "channel-141",
+    coinMinimalDenom: "uatom",
+  },
+  {
+    counterpartyChainId: "columbus-5",
+    sourceChannelId: "channel-72",
+    destChannelId: "channel-1",
+    coinMinimalDenom: "uluna",
+  },
+  {
+    counterpartyChainId: "crypto-org-chain-mainnet-1",
+    sourceChannelId: "channel-5",
+    destChannelId: "channel-10",
+    coinMinimalDenom: "basecro",
+  },
+  {
+    counterpartyChainId: "columbus-5",
+    sourceChannelId: "channel-72",
+    destChannelId: "channel-1",
+    coinMinimalDenom: "uusd",
+  },
+  {
+    counterpartyChainId: "secret-4",
+    sourceChannelId: "channel-88",
+    destChannelId: "channel-1",
+    coinMinimalDenom: "uscrt",
+  },
+  {
+    counterpartyChainId: "juno-1",
+    sourceChannelId: "channel-42",
+    destChannelId: "channel-0",
+    coinMinimalDenom: "ujuno",
+  },
+  {
+    counterpartyChainId: "juno-1",
+    sourceChannelId: "channel-169",
+    destChannelId: "channel-47",
+    coinMinimalDenom:
+      "cw20:juno168ctmpyppk90d34p3jjy658zf5a5l3w8wk35wht6ccqj4mr0yv8s4j5awr",
+    ics20ContractAddress:
+      "juno1v4887y83d6g28puzvt8cl0f3cdhd3y6y9mpysnsp3k8krdm7l6jqgm0rkn",
+  },
+  {
+    counterpartyChainId: "core-1",
+    sourceChannelId: "channel-4",
+    destChannelId: "channel-6",
+    coinMinimalDenom: "uxprt",
+  },
+  {
+    counterpartyChainId: "core-1",
+    sourceChannelId: "channel-4",
+    destChannelId: "channel-6",
+    coinMinimalDenom:
+      "ibc/A6E3AF63B3C906416A9AF7A556C59EA4BD50E617EFFE6299B99700CCB780E444",
+    ibcTransferPathDenom:
+      "transfer/channel-38/gravity0xfB5c6815cA3AC72Ce9F5006869AE67f18bF77006",
+  },
+  {
+    counterpartyChainId: "columbus-5",
+    sourceChannelId: "channel-72",
+    destChannelId: "channel-1",
+    coinMinimalDenom: "ukrw",
+  },
+  {
+    counterpartyChainId: "akashnet-2",
+    sourceChannelId: "channel-1",
+    destChannelId: "channel-9",
+    coinMinimalDenom: "uakt",
+  },
+  {
+    counterpartyChainId: "regen-1",
+    sourceChannelId: "channel-8",
+    destChannelId: "channel-1",
+    coinMinimalDenom: "uregen",
+  },
+  {
+    counterpartyChainId: "sentinelhub-2",
+    sourceChannelId: "channel-2",
+    destChannelId: "channel-0",
+    coinMinimalDenom: "udvpn",
+  },
+  {
+    counterpartyChainId: "irishub-1",
+    sourceChannelId: "channel-6",
+    destChannelId: "channel-3",
+    coinMinimalDenom: "uiris",
+  },
+  {
+    counterpartyChainId: "iov-mainnet-ibc",
+    sourceChannelId: "channel-15",
+    destChannelId: "channel-2",
+    coinMinimalDenom: "uiov",
+  },
+  {
+    counterpartyChainId: "emoney-3",
+    sourceChannelId: "channel-37",
+    destChannelId: "channel-0",
+    coinMinimalDenom: "ungm",
+  },
+  {
+    counterpartyChainId: "emoney-3",
+    sourceChannelId: "channel-37",
+    destChannelId: "channel-0",
+    coinMinimalDenom: "eeur",
+  },
+  {
+    counterpartyChainId: "microtick-1",
+    sourceChannelId: "channel-39",
+    destChannelId: "channel-16",
+    coinMinimalDenom: "utick",
+  },
+  {
+    counterpartyChainId: "likecoin-mainnet-2",
+    sourceChannelId: "channel-53",
+    destChannelId: "channel-3",
+    coinMinimalDenom: "nanolike",
+  },
+  {
+    counterpartyChainId: "impacthub-3",
+    sourceChannelId: "channel-38",
+    destChannelId: "channel-4",
+    coinMinimalDenom: "uixo",
+  },
+  {
+    counterpartyChainId: "bitcanna-1",
+    sourceChannelId: "channel-51",
+    destChannelId: "channel-1",
+    coinMinimalDenom: "ubcna",
+  },
+  {
+    counterpartyChainId: "bitsong-2b",
+    sourceChannelId: "channel-73",
+    destChannelId: "channel-0",
+    coinMinimalDenom: "ubtsg",
+  },
+  {
+    counterpartyChainId: "kichain-2",
+    sourceChannelId: "channel-77",
+    destChannelId: "channel-0",
+    coinMinimalDenom: "uxki",
+  },
+  {
+    counterpartyChainId: "panacea-3",
+    sourceChannelId: "channel-82",
+    destChannelId: "channel-1",
+    coinMinimalDenom: "umed",
+  },
+  {
+    counterpartyChainId: "bostrom",
+    sourceChannelId: "channel-95",
+    destChannelId: "channel-2",
+    coinMinimalDenom: "boot",
+  },
+  {
+    counterpartyChainId: "comdex-1",
+    sourceChannelId: "channel-87",
+    destChannelId: "channel-1",
+    coinMinimalDenom: "ucmdx",
+  },
+  {
+    counterpartyChainId: "cheqd-mainnet-1",
+    sourceChannelId: "channel-108",
+    destChannelId: "channel-0",
+    coinMinimalDenom: "ncheq",
+  },
+  {
+    counterpartyChainId: "stargaze-1",
+    sourceChannelId: "channel-75",
+    destChannelId: "channel-0",
+    coinMinimalDenom: "ustars",
+  },
+  {
+    counterpartyChainId: "chihuahua-1",
+    sourceChannelId: "channel-113",
+    destChannelId: "channel-7",
+    coinMinimalDenom: "uhuahua",
+  },
+  {
+    counterpartyChainId: "lum-network-1",
+    sourceChannelId: "channel-115",
+    destChannelId: "channel-3",
+    coinMinimalDenom: "ulum",
+  },
+];
