+++ packages/web/config/index.ts
@@ -1,0 +1,10 @@
+/** Config for chains to be approved by Keplr.
+ *  Includes name, currencies, bech32 info, and node endpoints.
+ */
+export * from "./chain-infos";
+
+/** Config for toggling the display of certain UI elements. */
+export * from "./feature-flag";
+
+/** Config for displaying IBC assets in a human-friendly way. */
+export * from "./ibc-assets";
